chech for root

+++ scripts/change_bin.sh
@@ -2,6 +2,11 @@
 
 echo -e "\n[Running change_bin] Deleting telnet, nc; changing to curlbk, wgetbk..."
 
+if [[ $EUID -ne 0 ]]; then
+   echo "This script must be run as root" 
+   exit 1
+fi
+
 rm $(which telnet) 2>/dev/null
 rm $(which nc) 2>/dev/null
